function/requirements.txt
azure-common==1.1.28
azure-communication-email==1.0.0
azure-core==1.30.2
azure-functions==1.20.0
azure-mgmt-core==1.4.0
certifi==2024.6.2
charset-normalizer==3.3.2
idna==3.7
isodate==0.6.1
msrest==0.7.1
oauthlib==3.2.2
psycopg2-binary==2.9.9
requests==2.32.3
requests-oauthlib==2.0.0
six==1.16.0
typing_extensions==4.12.2
urllib3==2.2.2
